Move generateBindingContext to detekt-parser
